backend/pyproject.toml
[tool.poetry]
name = "travel-assistant-backend"
version = "0.1.0"
description = ""
authors = ["AshwinHegde <oscii.ashwin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.10.6"
pydantic-ai = "^0.2.1"
python-dotenv = "^1.0.0"
nova-act = "1.0.2579.0"
playwright = "^1.40.0"
httpx = "^0.27.0"
tomli = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
